Learning HTTP In pubspec.yaml, 1) Added the dependency for HTTP.
In main.dart,
1) Changed the initialRoute from Home to Loading.

In choose_location.dart,
1) Cut the getData() and initState() functions and paste them in the Loading() class in loading.dart.

In loading.dart,
1) Imported the following packages:
-> http/http.dart to get the HTTP functions and variables.
-> dart:convert to convert pre-defined return types to favorable. In this instance, JSON is decoded to Map.
2) In the getData(), removed the Asynchronous Function Simulations. Received a Response type from the get() method with a dummy JSON Placeholder API.
We notice that we are getting the data of the type JSON. For our convenience, we decode the JSON body from the response using jsonDecode(response.body) and store the result as a Map object. As a result, we are now able to access each row of the Map and its contents.