Using a WorldTime API
In loading.dart,
1) Removed the simulation and the placeholder HTTP commands.
2) Changed the function's name from getData() to getTime() without changing its type.
3) Performed the HTTP action in three steps:
-> Made the request
Received a Response type after awaiting a get request from an API which has the time in London. Decoded the JSON to a Map type.
-> Got properties from the data
Extracted the relevant fields of data from the Map object.
-> Created a DateTime object
In order to parse the datetime extracted from the API, an instance of DateTime was used.